[#5228] Asynchronously update_quota_usage without using irods datazone
